forklift/requirements.txt
google-cloud-logging
google-cloud-storage
google-cloud-pubsub
